uHAL : Clean up new framework classes & functions for tests; refs #60
