+++ uhal/tests/src/common/test_docu_examples.cpp
@@ -33,6 +33,8 @@
 */
 
 #include "uhal/uhal.hpp"
+#include "uhal/tests/definitions.hpp"
+#include "uhal/tests/fixtures.hpp"
 #include "uhal/tests/tools.hpp"
 
 #include <boost/test/unit_test.hpp>
@@ -47,7 +49,7 @@ namespace tests {
 
 UHAL_TESTS_DEFINE_CLIENT_TEST_CASES(AddrTableDocuExamplesTestSuite, test_docu_addr_table_examples, DummyHardwareFixture,
 {
-  ConnectionManager manager ( sConnectionFile );
+  ConnectionManager manager ( connectionFileURI );
   // This line is majority of the test (i.e. load the addr table without exception).
   HwInterface hw=manager.getDevice ( "dummy.docu" + deviceId.substr(5) );
   // ***** "Single Register Address Table" example *****
